requirements.txt
daiquiri
tensorflow-gpu
daiquiri
tf-slim
